+++ docs/changelog.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Write docs for system identifier set/get/find/resolve sc-memory API
+- Extend set/get/find/resolve sc-memory API with out system identifier fiver parameters
 - Insert sc-keynodes in global knowledge base during sc-memory initiation 
 - CI for documentation build
 - Fix memory clear parameter reading from config file
@@ -27,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `SCs-helper` doesn't return generated identifier links ([207](https://github.com/ostis-ai/sc-machine/issues/207))
 - No json string message parsing in `sc-server` ([190](https://github.com/ostis-ai/sc-machine/issues/190))
 - `SCs-parser` create structure twice
 - `SCs-parser` doesn't unite all generated elements into structure ([205](https://github.com/ostis-ai/sc-machine/issues/205))